+++ NEWS.md
@@ -19,9 +19,17 @@ the dosing including dose amount and route.
   an error (fix #238)
 * Many new `assert_*` functions were added to standardize input checking in the
   style of the `checkmate` library.
-  
+
+## Bugs fixed
+
+* `superpostion()` now respects the interpolation and extrapolation types rather
+  than using default
+
 ## Breaking changes
 
+* The arguments `interp.method` and `extrap.method` have been replaced with
+  `method` and `auc.type` in the `interp.extrap.conc()` family of functions for
+  consistency with the rest of PKNCA (fix #244)
 * The AIC.list() function is no longer exported (it was never intended to be an
   external function).
 * The `depends` argument to `add.interval.col()` must either be NULL or a